Refactor User deletion section
